ページ内目次のマークアップを共通化
